# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API XCRM",
        "description": "Documentação da API do sistema XCRM",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/chamado-respostas": {
            "get": {
                "tags": [
                    "Chamado Respostas"
                ],
                "summary": "Listar todas as respostas dos chamados",
                "operationId": "1acc23f3aebed5082d18d9050e470660",
                "responses": {
                    "200": {
                        "description": "Lista de respostas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chamado Respostas"
                ],
                "summary": "Registrar nova resposta ao chamado",
                "operationId": "37a2e19c65db60b50c82fa556f43d22b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "chamado_id",
                                    "mensagem",
                                    "autor"
                                ],
                                "properties": {
                                    "chamado_id": {
                                        "type": "integer"
                                    },
                                    "mensagem": {
                                        "type": "string"
                                    },
                                    "autor": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "cliente"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resposta registrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/chamado-respostas/{id}": {
            "get": {
                "tags": [
                    "Chamado Respostas"
                ],
                "summary": "Exibir resposta específica",
                "operationId": "26f6e8a41e63948a7cb3e02c98640917",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da resposta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta encontrada",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resposta não encontrada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Chamado Respostas"
                ],
                "summary": "Atualizar resposta ao chamado",
                "operationId": "23f0e12d9eb237277438e094faee8a94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da resposta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mensagem": {
                                        "type": "string"
                                    },
                                    "autor": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "cliente"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resposta não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chamado Respostas"
                ],
                "summary": "Remover resposta de chamado",
                "operationId": "5dad4c2f7bfcbc14e735f25a0d341140",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da resposta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resposta removida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resposta não encontrada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Use \"Bearer {token}\" no campo de autorização",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}